requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
Pillow>=10.1.0
boto3>=1.34.0
django-cors-headers>=4.3.1
gunicorn>=21.2.0
celery>=5.3.4
redis>=5.0.1
django-redis>=5.4.0
django-ratelimit>=4.1.0
ffmpeg-python>=0.2.0
sentry-sdk>=1.39.0
django-storages>=1.14.2
twilio>=8.0.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
python-json-logger>=2.0.7
